+++ src/docker/events.go
@@ -140,6 +140,7 @@ func DebouncedExportDocker() {
 
 func onDockerStarted(containerID string) {
 	utils.Debug("onDockerStarted: " + containerID)
+	// exportService(containerID)
 	// BootstrapContainerFromTags(containerID)
 	DebouncedExportDocker()
 }