+++ changelog.md
@@ -3,6 +3,7 @@
  - Improved network IP resolution for containers, including supporting any network mode
  - Integrated MongoDB as container
  - Removed all sort of container bootstrapping (much faster boot)
+ - Cosmos now keep docker-compose files updated when updating containers, networks, etc...
 
 ## Version 0.13.0
  - Display container stacks as a group in the UI